Extract concern to manage outdated items
